death_note/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
capstone==4.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
enum34==1.1.6
filelock==3.0.10
idna==2.8
intervaltree==3.0.2
ipaddress==1.0.22
Mako==1.0.7
MarkupSafe==1.1.0
packaging==18.0
paramiko==2.4.2
pluggy==0.8.0
psutil==5.4.8
pwntools==3.12.1
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
pyelftools==0.25
Pygments==2.3.1
PyNaCl==1.3.0
pypandoc==1.4
pyparsing==2.3.0
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.5
requests==2.21.0
ROPGadget==5.4
six==1.12.0
sortedcontainers==1.5.10
toml==0.10.0
tox==3.6.0
unicorn==1.0.1
urllib3==1.24.1
virtualenv==16.1.0
